Updated submodules Fixed warning

+++ Sources/WaveDrivers/WaveDriverUSB.mm
@@ -256,7 +256,7 @@
 
 #pragma mark -
 
-- (NSUInteger) allowedChannels
+- (UInt16) allowedChannels
 {
     UInt16 channels;
